refactor: domain과 ui로 폴더 분리
* 리팩토링 요구사항
