+++ src/view/Validator.js
